working toward a locally runnable ddm platform impl
This is mostly for local testing and for the control plane team to use
for integration testing. What's kidna nice is this is mostly copy pasta
from the illumos DDM platform implementation. If this jsut works, this
local platform can be refactored into almost nothing and provide test
coverage for the production illumos platofrm.