add gateway to route add in dendrite handler
Originally, I thought just using the port index would be sufficient. And
it is to get the packet out the door, it's just missing the destination
mac ... :/. Adding the link local gateway let's the switch ASIC code
connect the dots between the egress spec and the available NDP table
state.