Toward a more monolithic routing platform.
The original goal of the Maghemite platform interface was to facilitate
the same routing algorithms on multiple underlying platforms in a
maintainable way. The motivation for doing this was to support both
vanilla-illumos hosts as well as sidecars.

However, the way things have taken shape with sidecar, and specifically
with dendrite obviate the need to maintain multiple platforms. The
dendrite stack presents the PCIe attached sidecar interfaces as regular
illumos interfaces, this means that in terms of communicating with
peers, a sidecar attached router looks exactly the same as a vanilla
illumos router. So the platform interface in this regard is a vacuous
abstraction. This is actually good, the purpose of the OS is to abstract
hardware, so let's let it do that.

There are differences in how the routing tables are managed between
dendrite and illumos. However, it makes much more sense to abstract a
small set of common functions in a tight interface rather than pretend
that these are completely different "platforms".

All the above is to say that at the present time the platform model is
getting in the way more that it is helping, and it's mostly a hammer in
search of a nail. We have no driving use case for any other platforms so
it does not make sense to continue with this design.

Another aspect of the first cut at a platform design that proved
problematic was the use of channels for communicating between the
platform and the higher level routing code. Some of the communications
between routers are most simply performed a synchronous request-reply
exchanges. Even through this is how the communications are implemented at
the platform level, the routing code does not see it this way because
there is a bidirectional channel between it and the platform. This turns
communication that is synchronous into locally asynchronous comms. And
while this certainly can be handled, it adds middleman buffering
complexity for no reason, adding cruft and making a larger bug surface.

This commit is the start of a more monolithic Maghemite design, geared
toward illumos - leaving hardware abstraction largely where it belongs,
in the OS.