ddmd working sans SIOCGNDS
SIOCGNDS was being used in transit mode when targeting the dendrite
substrate to connect the dots from gateway to interface. This commit
replaces this by plumbing the interface name through to the
advertisement handlers that create routes on the local system. With this
information the function that creates dendrite routes can drive the port
number from the tfport name.