Switch to IpAddr for identifying BGP session
Switch from SocketAddrPair to IpAddr for distinguishing between BGP
sessions. A BGP peer is identified by the IP we are connecting to,
and multiple TCP sessions via the same IP isn't something we're
looking to support. So instead of trying to handle all the cases
where the local SocketAddr may or may not be available, simplify
things by using the configured peer IP which will always be
present regardless of the peer's FSM State.

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>