Chore: Bump OPTE to `88adb1a`
Brings us up to OPTE v0.35.x, and is part of bringing tunnelled LSO into
use by omicron.

An extra point is that I needed to bump the rust version used in this
repo, since OPTE is on the 2024 rust edition. Renovate looks to have
rate-limited itself at some point, so this has languished for some time.