Use `oxide-tokio-rt` rather than `#[tokio::main]` (#523)
As described in [RFD 579], we are now recommending certain non-default
Tokio runtime configurations in all software we deploy in production.
The [`oxide-tokio-rt`] crate provides a way to easily get these common
recommended configurations by using its functions to initialize the
runtime rather than `#[tokio::main]`

Presently, `oxide-tokio-rt` does the following:

- Enables DTrace probes for Tokio runtime events using [`tokio-dtrace`]
- Disables the [LIFO slot optimization][lifo], which was the root cause
  of omicron#8334

But, we anticipate that if there are additional Tokio configurations
which we want to recommend in the future, we'll add them here as well.

This commit updates `dendrite` to initialize the Tokio runtime using
`oxide-tokio-rt`. I've added a Clippy config for warning on uses of
`#[tokio::main]`, as well, to avoid forgetting to use `oxide-tokio-rt`
if new binaries are added. All the configs we set currently in
`oxide-tokio-rt` requires the `tokio_unstable` RUSTFLAGS cfg, but it
looks like that's already being set, so that shouldn't be a problem.

I've changed every binary to use `oxide-tokio-rt` with the exception of
`mg-package` --- it didn't seem necessary to use the production-like
config there and maybe avoiding an additional dependency to fetch would
make it build slightly faster, although I'm unconvinced. I did change
the lab binaries to use `oxide-tokio-rt`, as I figured that the DTrace
probes could come in handy there.

[RFD 579]: https://rfd.shared.oxide.computer/rfd/0579
[`oxide-tokio-rt`]: https://github.com/oxidecomputer/oxide-tokio-rt
[`tokio-dtrace`]: https://github.com/oxidecomputer/tokio-dtrace
[lifo]:
    https://rfd.shared.oxide.computer/rfd/0579#_disabling_the_lifo_slot_optimization