bgp: align bestpath with RFC 4721
Adds the missing steps from the BGP-4 "Tie Breaking" algorithm as defined by
RFC 4721. This includes updates to some data structures
(mainly BgpPathProperties) in order to support the additional steps.
The one step from the RFC that was not included was the IGP cost calculation,
as maghemite does not support an IGP other than DDM, which is run in a separate
routing domain as BGP (BGP is strictly an EGP in this stack).

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>