remove invalid static prefixes from db on startup
Second half of the fix for #529. First we moved to a constructor that
enforced all host bits be zeroed when creating a Prefix* type. Now, we
check whether the routes in the on-disk db are valid (host bits unset)
and remove them if they're invalid.
This also moves most of the logic into methods of the Prefix* types,
with the exception of the mg-admin-client mirrors of them, as progenitor
can't know about methods/traits from the openapi spec so that logic
has to live somewhere.

Fixes: #529

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>