bgp: fix IdleHoldTimer handling
The idle hold timer was not being handled properly before. This led to
situations where a session that was configured as passive to transition
from the Idle FSM state into Connect, when passive sessions should only
ever go into Active.  This adds explicit handling of the timer in the
FSM when transitions to Idle state are occurring or when the
IdleHoldTimerExpires FSM Event is received.

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>