bgp: refactor tests, bind before TCP connect()
Refactor tests to only use a topology setup function that supports an
arbitrary number of routers.

In tests, bind() socket before calling connect() for TCP connections to
ensure Source IP is not selected by the OS. This prevents a logical
router from opening a TCP connection from a Source IP other than what
it's listening on.
e.g.
 - r1 listens on 127.0.0.1, connects to r2 via 127.0.0.2
 - r2 listens on 127.0.0.2, connects to r1 via 127.0.0.1

Without this change, the SIP is selected by the OS and r1 could attempt
to connect to r2 with SIP=127.0.0.2. If this happens, r2 denies the
connection because it only knows about the IP it has configured as a
peer, i.e. 127.0.0.1.

With this change, r1 binds the socket to 127.0.0.1 before connecting to
127.0.0.2 and r2 sees an inbound request from the peer's configured IP.

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>