bgp: start Cnx recv loop after conn registration
Move the starting of the recv loop into its own method which gets called
by SessionRunner.register_conn(). This resolves a race condition where
it is possible for a new connection to recv() a Message and push it into
the SessionRunner's event queue before the "new connection" event with
the handle for the new connection is pushed into the queue
i.e.
ConnectionEvent::Message could arrive before SessionEvent::Connected or
SessionEvent::TcpConnectionConfirmed.

Also update Active and Connect states to no longer handle Message events
as valid events. Since we don't implement DelayOpen, our immediate
reaction to a new connection is to send an Open and move into OpenSent
without reading from the event queue again until after the FSM
transition. Once (if) DelayOpen is fully implemented, Opens received
while DelayOpenTimer is running will need to be handled with a
reciprocal Open, a Keepalive, and an FSM transition into OpenConfirm.

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>