bgp: remove restart_connect_retry() method
RFC 4271 uses several unique terms when referring to how a timer should
be manipulated: "sets," "resets," "starts," "restarts," "clears,"
"stops," "sets to zero," and "sets to initial value." None of these have
a definition (e.g. what does zero mean in context of the timer? does the
timer count up from zero towards the interval, or does it count down
from the interval towards zero?) or any clarification as to whether the
different terms are intended to have differences in operation. However,
20+ year old IETF mailing list threads seem to have folks on both sides
("this is obvious, we can't explain English" vs "this is confusing and
adds friction"). Despite discussion around making the phrasing more
clear in the RFC, that was never done.

After more time and analysis of the FSM, it seems that the only places
where ConnectRetryTimer should be started is in Idle, Connect or Active,
and that the other places where it's manipulated are to stop/disable it.

After replacing restart_connect_retry() calls with direct logic to stop
ConnectRetryTimer the only calls that were left were places where we
already had a passive connection check, so there was very little value
still being gained from the wrapper.

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>