bgp: fix ConnectRetryTimerExpires deadlock
In Connect we start off by immediately locking self.session before
entering the event handler loop, and we never drop the lock. If the
initial connect fails, we restart ConnectRetryTimer and continue the
loop. When the ConnectRetryTimerExpires event is received, we try to
lock self.session a second time, resulting in a deadlock. This fixes
that issue by moving the self.session lock and the connect calls into
braces that force the lock to be dropped when things are done.

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>