bgp: fix collision resolution semantics
Connection Collision Resolution requires that the BGP-ID is used as a
tie-breaker to determine which connection to close: the connection
initiated by the router with the lowest ID is closed. We were correctly
using the BGP-ID as the tie-breaker, but were implicitly using the "new"
connection (whichever completed second) as if it were always initiated
by the peer and the "existing" connection (whichever completed first)
as if it were initiated by the local system. This is a bad assumption
since the ordering of connection completion is non-deterministic. This
fixes issues that can arise when those assumptions are made,
i.e.
r1 and r2 agree that r2 has the higher BGP-ID, but each decide to close
a different connection.

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>