bgp: fix IPv6 next-hops sent for IPv4 updates
The accept() path (via the Dispatcher) for TCP was incorrectly passing
the local bind address to the BgpConnection constructor (with_conn())
which is set to [::]:179 in non-test environments, which resulted in
us later encoding this into an Update message in the NEXT_HOP field.
This addresses the issue by correctly passing the local SocketAddr of
the live socket, which is an IPv4 address for an IPv4 TCP connection.
This also adds both tx and rx validation of NEXT_HOP length based on AFI
of prefixes and their next-hop.

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>