bgp, mgd: more bind() and connect() fixes
Updates connect() logic to use either a plain IPv4 or IPv6 address from
the new TcpStream's local SocketAddr, rather than using it as-is. This
is needed because our listening socket is a dual-stack socket (binding
to [::] with `v6_only` disabled), so any new IPv4 TcpStreams we accept()
from it are presented to the application as an IPv6 socket with an
IPv4-mapped IPv6 address (e.g. a new connection from 192.168.0.1 is
handed to the application with a peer IP of ::ffff:192.168.0.1).

The IP we strip from the TcpStream is then used if we attempt to bind()
the socket to a specific Source IP before initiating an outbound
connection, which we bail out of upon a failure. So no outbound
connections will be initiated while this is busted. The IP is also used
to determine the address-family of the next-hop we encode in sent Update
messages, so we could inadvertently encode 16 bytes instead of 4 for the
next-hop if this is wrong.

Updates the bind_addr of a SessionRunner to be a Option all the way
up through the API handler (and the test instantiation), where it is
always set to None by the API (until we add support for update-source).
This prevents erroneous bind() calls in non-test environments.

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>