bgp: reject second connection with same initiator
Add a ConnectionCreator check in OpenSent/OpenConfirm to ensure we never
enter FsmState::ConnectionCollision with two connections that were
initiated by the same peer (local or remote).

It was observed during manual stress testing that mgd would sometimes
enter Connection Collision with two connections that it initiated. This
happens because connect() calls for outbound TCP sockets are currently
handled as a one-shot operation in a separate thread without any
synchronization methods to allow for a graceful shutdown to occur (e.g.
sending a Shutdown message to the connect thread via a channel).
However, since we use connect_timeout(), the timing window for these
duplicate outbound connections is bounded by the timeout supplied
(currently 100ms). Regardless, any timing window is still a window,
even if this was only observed irregularly (seen 74 times after running
overnight with 2 peers who were resetting the session once per second),
so this adds logic to disallow this condition. If a second connection
completes, the FSM will check its direction (inbound or outbound) and
reject it if it matches the existing connection's direction.

This also removes a couple (now) unneeded clippy lints, and renames
SessionEvent::Connected to SessionEvent::TcpConnectionAcked (now
aligning with RFC event 16, Tcp_CR_Acked).

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>