bgp: refactor resolve_collision()
Renames PrimaryConnection enum to ConnectionKind.

Updates resolve_collision() to take BGP-ID as an argument rather than
taking it from one of the connections, and to use ConnectionKind enum
for the two connections instead of strictly PeerConnections. This
allows the callers to consistently order their args by connection
completion time (`exist` = first, `new` = second) regardless of how far
along the connection is.

This is preparation for a refactor of the connection collision logic
where we enter resolve_collision() earlier than we did before.

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>