bgp: disable IdleHoldTimer in test topologies
When both peers are configured to use identical IdleHoldTimer, they can
encounter a situation where one peer tries to connect to the other while
it's in Idle, which gets rejected. Upon that rejection (rather, the
resulting send() failure) the first peer drops back to Idle. Then when
the second peer comes out of Idle and attempts to connect, the first
peer will reject its connection while it's sitting in Idle. This death
spiral can sometimes happen indefinitely, which presents itself as
random test failures when the BGP session fails to establish within the
30 second timout in the tests. By transitioning immediately out of Idle
(i.e. disabling IdleHoldTime) without reading from the event queue, we
avoid the whole situation where connections are rejected by the rules of
the Idle FSM state.

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>