bgp: revert non-standard notification handling
In an attempt to avoid the death spiral involved with both peers using
IdleHoldTime (getting stuck in an Idle <-> Connect loop), I had tried
adding special handling for Collision Connection Resolution Notification
messages. After realizing that this problem is inherent to the Idle
state (rejecting connections that complete asynchronously) and adjusting
tests to disable IdleHoldTimer (aligning with the default config), I
realized that we actually do not need this non-compliant handling.

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>