bgp: initial cleanup/improvement of md5 code
This refactors some of the md5 code to reduce code reuse via helpers.
This also re-organizes the connection_tcp.rs file by grouping together
target-specific code.
The main functional change in this commit is that the connect() method
now does resource allocation/initialization before spawning the new
thread, giving it a chance to actually return errors to the caller.
std::thread::spawn is also replaced with std::thread::Builder::spawn, as
it returns Result::Err instead of panicking if the thread creation fails

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>