bgp, rdb: refactor Prefix validation
Move validation of Prefix content into methods in rdb::types. This
allows the consumer of a prefix to easily check if it is semantically
semantically valid for routing/RIB purposes while centralizing the logic
that makes this determination.

Remove new_unchecked() method of Prefix and Co.

Update a couple type signatures to take slices instead of owned Vecs.

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>