bgp: mark event_rx channel disconnected unreachable
It is currently impossible for the receiver of the FSM event queue
(implemented via mpsc channel) to return RecvTimeoutError::Disconnected
because there is always a clone of its sender until both the sender and
receiver go out of scope (when a peer is removed and its SessionRunner
is dropped) at the same time. It would be mighty difficult to imagine a
valid scenario where all clones of the sender get dropped when one of
the clones has the same owner as the receiver.

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>