bgp: defensive handling of FSM timer events
React to unexpected timer events more defensively. While the RFC says in
the FSM description to respond to unexpected timers by transitioning to
Idle and sending an FSM Error Notification, many of these timer events
are both internal and harmless. Instead of closing the connection and
restarting the FSM from Idle, stop the offending timer, log it as a bug,
and continue where the FSM left off.

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>