bgp: fix first batch of thread leaks
Fixes for 3 different thread leaks:
1: MD5 SA thread not closed when BgpConnection is dropped
2: ConnectionClock thread not closed when BgpConnection is dropped
3: SessionClock thread not closed when SessionRunner is dropped

All 3 of these leaks come down to the same fundamental issue:
The Drop impl of the type responsible for these threads was incorrectly
checking if Arc::strong_count == 1 before setting the shutdown/dropped
flag (Arc<AtomicBool>) to true. The correct behavior is for the Drop
impl of the owning types to unconditionally set the dropped flag true.

Context:
The initial implementation of the SessionRunner connection registry used
a BTreeMap that held clones of BgpConnections. These BgpConnections hold
an Arc<AtomicBool> that is used to signal shutdown to its child threads
(MD5 SA keepalive thread + ConnectionClock). Because there were multiple
BgpConnection clones that each had an Arc clone of the AtomicBool, we
needed to check whether the strong_count had reached 1 so we didn't set
the dropped flag too early.
After the recent rewrite of the ConnectionRegistry, BgpConnections are
no longer cloned. Instead, there is proper shared ownership via Arc
wrappers (Arc<BgpConnection>). Since there is now only a single instance
of the BgpConnection, it is safe (and necessary) for the dropped flag to
be set by the Drop impl without checking the strong count of the Arc
wrapping the dropped flag.

The ConnectionClock and SessionClock types also followed this pattern
(which was never needed for the SessionClock), causing both of these
threads to get leaked during shutdown as well. The SessionClock leak
only occurred when the entire SessionRunner was dropped, i.e. when a
neighbor is unconfigured.

Along with the corrected Drop impl, this moves the ThreadState enum
into mg-common and provides a ManagedThread wrapper which holds both a
dropped flag and a ThreadState enum.

This also adds a testing utility to check the number of threads in the
current process and a bgp test that ensures there are no threads leaked
by the lifecycle of the BgpConnection, SessionRunner or Router types.

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>