bgp: add explicit router lifecycle
The Router type has a shutdown() method and a shutdown flag but the
behavior didn't seem to be fully functional. The shutdown flag was only
ever written, but never read (or cloned).

This introduces explicit handling for the shutdown flag:
- shutdown() sets Router.shutdown flag and sets shutdown for sessions
- run() unsets Router.shutdown flag and starts sessions
- admin events now check Router.shutdown before attempting to propagate
  changes to neighbors (e.g. when updating originated routes)

Additionally, a Drop impl is added for Router which calls shutdown on
all sessions so their threads actually close when a Router is deleted.

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>