Use thread names and fix bgp test thread counter
Move to using std::thread::Builder for all BGP threads instead of
calling std::thread::spawn directly. This allows us to name our threads
and gives us the opportunity to handle a spawn failure rather than
panicking unconditionally if the spawn goes wrong.
Additionally, the logic of the test functions that counted the number of
active threads needed to be tweaked, as it checked the total number of
threads for this process rather than looking just for BGP threads. This
caused issues in the thread count because it didn't filter out unrelated
threads threads like slog-async, sled-io, or the rdb::Db reaper.

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>