bgp: test threads serially, fix fanout on restart
Pull in the serial_test dependency so we can make sure the new thread
count tests are not run in parallel with other tests that could spawn
bgp threads. In addition to this, have a wait condition for the baseline
number of bgp threads to be 0, just in case tests are not run using
nextest (`cargo test` executes all tests in the same process).

In the Router code, move SessionRunner thread spawning into its own
method so we avoid the possibility of having subtly different code to do
the same thing. Also make sure that we update fanout during shutdown()
and run() to support a Router being restarted.

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>